requirements.txt
django==4.2.4
djangorestframework==3.14.0
redis==4.6.0
psycopg2-binary==2.9.6
pillow==10.0.0
django-cors-headers==4.2.0
cryptography==3.4.8